+++ src/test/java/org/example/ForFibonacciTest.java
@@ -2,6 +2,7 @@ package org.example;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -18,7 +19,7 @@ class ForFibonacciTest extends SystemOverride {
 
     @Test
     public void getFibonacci() {
-        List<Integer> fibonacci = forFibonacci.getFibonacci(10);
+        List<BigInteger> fibonacci = forFibonacci.getFibonacci(10);
 
         assertEquals(10, fibonacci.size());
         assertEquals("[0, 1, 1, 2, 3, 5, 8, 13, 21, 34]", fibonacci.toString());
