+++ src/test/java/org/example/RecursiveFibonacciTest.java
@@ -2,9 +2,10 @@ package org.example;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigInteger;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class RecursiveFibonacciTest extends SystemOverride {
 
@@ -18,7 +19,7 @@ class RecursiveFibonacciTest extends SystemOverride {
 
     @Test
     public void getFibonacci() {
-        List<Integer> fibonacci = recursiveFibonacci.getFibonacci(10);
+        List<BigInteger> fibonacci = recursiveFibonacci.getFibonacci(10);
 
         assertEquals(10, fibonacci.size());
         assertEquals("[0, 1, 1, 2, 3, 5, 8, 13, 21, 34]", fibonacci.toString());
